complete refactoring changes for new utils organization
